requirements.txt
pytest==6.2.4
PyYAML==5.4.1
numpy==1.19.5
pytz==2022.1
pre-commit==2.21.0
